harness: extract invoke context callback into callback mod
