+++ harness/src/callback/mod.rs
@@ -2,3 +2,4 @@
 
 #[cfg(feature = "invocation-inspect-callback")]
 pub mod invocation_inspect;
+pub mod invoke_context;
